refactor: navigation tabs sidebar
